+++ dzgui/const/boilerplate.py
@@ -1,0 +1,16 @@
+"""
+Generic defaults used when initializing a config file from scratch
+"""
+config_boilerplate = {
+    "bm_api": "",
+    "fav_server": "",
+    "fav_label": "",
+    "name": "",
+    "fullscreen": False,
+    "steam_api": "",
+    "default_steam_path": "",
+    "client": "",
+    "ip_list": [],
+    "use_miles": False,
+    "start_tab": 0,
+}